requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
selenium==4.11.2
PyChromeDevTools==0.4
psutil==5.9.5
python-dotenv==1.0.0
pymysql==1.1.0